+++ astropy/io/ascii/fastbasic.py
@@ -58,6 +58,59 @@ class FastBasic(object):
         # can be overridden for specialized headers
         self.engine.read_header()
 
+    @staticmethod
+    def _get_col_type(col_dtype):
+        # get the column type as Type classes defined in core
+        col_dtype = str(col_dtype)
+        if 'int' in col_dtype:
+            col_type = core.IntType
+        elif 'float' in col_dtype:
+            col_type = core.FloatType
+        elif 'bool' in col_dtype:
+            col_type = core.BoolType
+        elif 'str' in col_dtype:
+            col_type = core.StrType
+        else:
+            col_type = core.AllType
+        return col_type
+
+    def _convert_vals(self, data):
+        converter_first = None
+        modified_col = None
+
+        if self.converters is not None:
+            for col_name in data:
+                converters = self.converters.get(col_name, None)
+                if converters is not None:
+                    if len(converters) == 0:
+                        raise ValueError('Column {} failed to convert: no converters defined'.format(col_name))
+                    col_dtype = data[col_name].dtype
+                    col_type = self._get_col_type(col_dtype)
+
+                    try:
+                        for converter in converters:
+                            converter_func , converter_type = converter
+                            if not issubclass(converter_type, core.NoType):
+                                raise ValueError()
+                            if converter_first is None:
+                                converter_first = (converter_func , converter_type)
+                            if issubclass(converter_type, col_type):
+                                data[col_name] = converter_func(data[col_name], fast=True)
+                                modified_col = True
+                                break
+
+                    except (ValueError, TypeError):
+                        raise ValueError('Error: invalid format for converters, see documentation\n{}'.format(converters))
+
+                    if modified_col is None:
+                        converter_func , converter_type = converter_first
+                        data[col_name] = converter_func(data[col_name], fast=True)
+
+                    modified_col = None
+                    converter_first = None
+
+        return data
+
     def read(self, table):
         """
         Read input data (file-like object, filename, list of strings, or
@@ -77,15 +130,14 @@ class FastBasic(object):
             raise core.ParameterError("The C reader only supports 1-char delimiters")
         elif len(self.quotechar) != 1:
             raise core.ParameterError("The C reader only supports a length-1 quote character")
-        elif 'converters' in self.kwargs:
-            raise core.ParameterError("The C reader does not support passing "
-                                      "specialized converters")
         elif 'Outputter' in self.kwargs:
             raise core.ParameterError("The C reader does not use the Outputter parameter")
         elif 'Inputter' in self.kwargs:
             raise core.ParameterError("The C reader does not use the Inputter parameter")
         elif 'data_Splitter' in self.kwargs or 'header_Splitter' in self.kwargs:
             raise core.ParameterError("The C reader does not use a Splitter class")
+
+        self.converters = self.kwargs.pop('converters', None)
 
         self.strict_names = self.kwargs.pop('strict_names', False)
 
@@ -109,10 +161,12 @@ class FastBasic(object):
 
         with set_locale('C'):
             data, comments = self.engine.read(try_int, try_float, try_string)
-
         meta = OrderedDict()
         if comments:
             meta['comments'] = comments
+
+        data = self._convert_vals(data)
+
         return Table(data, names=list(self.engine.get_names()), meta=meta)
 
     def check_header(self):