+++ CHANGES.rst
@@ -21,6 +21,7 @@ New Features
 - ``astropy.cosmology``
 
 - ``astropy.io.ascii``
+  - Added ``converters`` parameter to ``FastBasic.read()``, to optimize memory usage. [#5885].
 
 - ``astropy.io.fits``